models/smilesdft_clip/requirements.txt
transformers 
torch-optimizer
datasets
scikit-learn
scipy>=1.12.0
numpy==1.26.4
pandas>=1.4.0
tqdm>=4.66.4
rdkit>=2024.3.5
imageio==2.34.1
hydra-core==1.3.2
omegaconf==2.3.0
pdbpp==0.10.2
torchvision==0.18.0
tqdm==4.66.4
requests>=2.32.0
opencv-python>=4.11.0.86